Reorganize documentation for Registration Coordinator
This consolidates information spread across the repo's README and the
nested docs around setting up registration. A lot of the latter is
outdated anyway. As part of this cleanup, I also tried to rewrite the
instructions such that they're not always Eventbrite-specific. There are
just some things we should try to do even if we're using a different
ticketing platform.